Some cleanup, renames
