+++ open_instruct/rlvr_curriculum.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 
 import math
-from dataclasses import dataclass
-from typing import Any
+from collections.abc import Callable
+from dataclasses import dataclass, field
+from typing import Any, Literal
 
 import numpy as np
 import torch
@@ -55,45 +56,40 @@ def _normalize_probs(values: np.ndarray) -> np.ndarray:
 
 
 @dataclass
-class DifficultyCurriculumConfig:
-    enabled: bool = False
-    difficulty_field: str = "difficulty"
+class DifficultyCurriculumMetadataConfig:
+    field: str = "difficulty"
     posterior_mean_field: str = "posterior_mean"
     bucket_index_field: str = "bucket_index"
     bucket_count_field: str = "bucket_count"
-    easy_focus_steps: int = 100
-    bootstrap_target_bucket_ratio: float = 0.125
-    warmup_target_bucket_ratio: float = 0.5
-    final_target_bucket_ratio: float = 1.0
+    strict: bool = True
+
+
+@dataclass
+class DifficultyCurriculumScheduleConfig:
+    bootstrap_steps: int = 100
     warmup_steps: int = 500
-    total_curriculum_steps: int = 10000
+    total_steps: int = 10000
+    bootstrap_target: float = 0.125
+    warmup_target: float = 0.5
+    final_target: float = 1.0
     min_hard_frac: float = 0.05
     max_hard_frac: float = 0.50
     bucket_sigma: float = 0.0
-    easy_focus_sigma: float = 0.0
-    uncertainty_weight: float = 0.5
-    adaptive_enabled: bool = False
-    adaptive_update_every: int = 50
-    adaptive_learning_signal_weight: float = 0.7
-    adaptive_exploration_weight: float = 0.3
-    adaptive_blend_weight: float = 0.5
-    seed: int = 0
-    strict_metadata: bool = True
-    epsilon: float = 1e-8
+    bootstrap_sigma: float = 0.0
 
     def __post_init__(self) -> None:
-        if self.easy_focus_steps < 0:
-            raise ValueError("easy_focus_steps must be >= 0")
-        if not 0.0 <= self.bootstrap_target_bucket_ratio <= 1.0:
-            raise ValueError("bootstrap_target_bucket_ratio must be in [0, 1]")
-        if not 0.0 <= self.warmup_target_bucket_ratio <= 1.0:
-            raise ValueError("warmup_target_bucket_ratio must be in [0, 1]")
-        if not 0.0 <= self.final_target_bucket_ratio <= 1.0:
-            raise ValueError("final_target_bucket_ratio must be in [0, 1]")
+        if self.bootstrap_steps < 0:
+            raise ValueError("bootstrap_steps must be >= 0")
         if self.warmup_steps < 0:
             raise ValueError("warmup_steps must be >= 0")
-        if self.total_curriculum_steps <= 0:
-            raise ValueError("total_curriculum_steps must be > 0")
+        if self.total_steps <= 0:
+            raise ValueError("total_steps must be > 0")
+        if not 0.0 <= self.bootstrap_target <= 1.0:
+            raise ValueError("bootstrap_target must be in [0, 1]")
+        if not 0.0 <= self.warmup_target <= 1.0:
+            raise ValueError("warmup_target must be in [0, 1]")
+        if not 0.0 <= self.final_target <= 1.0:
+            raise ValueError("final_target must be in [0, 1]")
         if not 0.0 <= self.min_hard_frac <= 1.0:
             raise ValueError("min_hard_frac must be in [0, 1]")
         if not 0.0 <= self.max_hard_frac <= 1.0:
@@ -102,20 +98,106 @@ class DifficultyCurriculumConfig:
             raise ValueError("min_hard_frac must be <= max_hard_frac")
         if self.bucket_sigma < 0:
             raise ValueError("bucket_sigma must be >= 0")
-        if self.easy_focus_sigma < 0:
-            raise ValueError("easy_focus_sigma must be >= 0")
+        if self.bootstrap_sigma < 0:
+            raise ValueError("bootstrap_sigma must be >= 0")
+
+
+@dataclass
+class DifficultyCurriculumAdaptiveConfig:
+    enabled: bool = False
+    update_every: int = 50
+    learning_weight: float = 0.7
+    exploration_weight: float = 0.3
+    blend: float = 0.5
+
+    def __post_init__(self) -> None:
+        if self.update_every <= 0:
+            raise ValueError("update_every must be > 0")
+        if not 0.0 <= self.learning_weight <= 1.0:
+            raise ValueError("learning_weight must be in [0, 1]")
+        if not 0.0 <= self.exploration_weight <= 1.0:
+            raise ValueError("exploration_weight must be in [0, 1]")
+        if not 0.0 <= self.blend <= 1.0:
+            raise ValueError("blend must be in [0, 1]")
+
+
+@dataclass
+class DifficultyCurriculumConfig:
+    metadata: DifficultyCurriculumMetadataConfig = field(default_factory=DifficultyCurriculumMetadataConfig)
+    schedule: DifficultyCurriculumScheduleConfig = field(default_factory=DifficultyCurriculumScheduleConfig)
+    adaptive: DifficultyCurriculumAdaptiveConfig = field(default_factory=DifficultyCurriculumAdaptiveConfig)
+    uncertainty_weight: float = 0.5
+    seed: int = 0
+    epsilon: float = 1e-8
+
+    def __post_init__(self) -> None:
         if not 0.0 <= self.uncertainty_weight <= 1.0:
             raise ValueError("uncertainty_weight must be in [0, 1]")
-        if self.adaptive_update_every <= 0:
-            raise ValueError("adaptive_update_every must be > 0")
-        if not 0.0 <= self.adaptive_learning_signal_weight <= 1.0:
-            raise ValueError("adaptive_learning_signal_weight must be in [0, 1]")
-        if not 0.0 <= self.adaptive_exploration_weight <= 1.0:
-            raise ValueError("adaptive_exploration_weight must be in [0, 1]")
-        if not 0.0 <= self.adaptive_blend_weight <= 1.0:
-            raise ValueError("adaptive_blend_weight must be in [0, 1]")
         if self.epsilon <= 0:
             raise ValueError("epsilon must be > 0")
+
+
+@dataclass
+class DifficultyCurriculumArgs:
+    curriculum: Literal["none", "difficulty"] = "none"
+    curriculum_metadata_field: str = "difficulty"
+    curriculum_posterior_mean_field: str = "posterior_mean"
+    curriculum_bucket_index_field: str = "bucket_index"
+    curriculum_bucket_count_field: str = "bucket_count"
+    curriculum_strict_metadata: bool = True
+    curriculum_bootstrap_steps: int = 100
+    curriculum_warmup_steps: int = 500
+    curriculum_total_steps: int = 10000
+    curriculum_bootstrap_target: float = 0.125
+    curriculum_warmup_target: float = 0.5
+    curriculum_final_target: float = 1.0
+    curriculum_min_hard_frac: float = 0.05
+    curriculum_max_hard_frac: float = 0.50
+    curriculum_bucket_sigma: float = 0.0
+    curriculum_bootstrap_sigma: float = 0.0
+    curriculum_uncertainty_weight: float = 0.5
+    curriculum_adaptive: bool = False
+    curriculum_adaptive_update_every: int = 50
+    curriculum_adaptive_learning_weight: float = 0.7
+    curriculum_adaptive_exploration_weight: float = 0.3
+    curriculum_adaptive_blend: float = 0.5
+
+    def build_curriculum_config(self, *, seed: int) -> DifficultyCurriculumConfig | None:
+        if self.curriculum == "none":
+            return None
+        if self.curriculum != "difficulty":
+            raise ValueError(f"Unsupported curriculum type: {self.curriculum}")
+
+        return DifficultyCurriculumConfig(
+            metadata=DifficultyCurriculumMetadataConfig(
+                field=self.curriculum_metadata_field,
+                posterior_mean_field=self.curriculum_posterior_mean_field,
+                bucket_index_field=self.curriculum_bucket_index_field,
+                bucket_count_field=self.curriculum_bucket_count_field,
+                strict=self.curriculum_strict_metadata,
+            ),
+            schedule=DifficultyCurriculumScheduleConfig(
+                bootstrap_steps=self.curriculum_bootstrap_steps,
+                warmup_steps=self.curriculum_warmup_steps,
+                total_steps=self.curriculum_total_steps,
+                bootstrap_target=self.curriculum_bootstrap_target,
+                warmup_target=self.curriculum_warmup_target,
+                final_target=self.curriculum_final_target,
+                min_hard_frac=self.curriculum_min_hard_frac,
+                max_hard_frac=self.curriculum_max_hard_frac,
+                bucket_sigma=self.curriculum_bucket_sigma,
+                bootstrap_sigma=self.curriculum_bootstrap_sigma,
+            ),
+            adaptive=DifficultyCurriculumAdaptiveConfig(
+                enabled=self.curriculum_adaptive,
+                update_every=self.curriculum_adaptive_update_every,
+                learning_weight=self.curriculum_adaptive_learning_weight,
+                exploration_weight=self.curriculum_adaptive_exploration_weight,
+                blend=self.curriculum_adaptive_blend,
+            ),
+            uncertainty_weight=self.curriculum_uncertainty_weight,
+            seed=seed,
+        )
 
 
 class AdaptiveBucketStats:
@@ -244,214 +326,56 @@ class _ParsedDifficultyMetadata:
     error: str | None
 
 
-class BetaBinomialDifficultySampler(Sampler[int]):
-    """Bucket-aware curriculum sampler that uses beta-binomial difficulty metadata."""
+@dataclass(frozen=True)
+class _DifficultyBucketIndex:
+    index_to_bucket: dict[int, int]
+    bucket_to_indices: tuple[tuple[int, ...], ...]
+    bucket_weights: tuple[torch.Tensor, ...]
+    bucket_count: int
+    metadata_fallback_count: int
 
-    def __init__(self, dataset, num_samples: int, config: DifficultyCurriculumConfig, global_step_getter) -> None:
-        if num_samples <= 0:
-            raise ValueError("num_samples must be > 0")
 
-        self.dataset = dataset
-        self.num_samples = num_samples
+class _DifficultyCurriculumSchedule:
+    def __init__(self, config: DifficultyCurriculumScheduleConfig, bucket_count: int) -> None:
         self.config = config
-        self.global_step_getter = global_step_getter
+        self.bucket_count = bucket_count
 
-        self._generator = torch.Generator()
-        self._generator.manual_seed(self.config.seed)
-
-        self._excluded_indices: set[int] = set()
-        self._index_to_bucket: dict[int, int] = {}
-        self.bucket_count = 1
-        self.metadata_fallback_count = 0
-
-        self._base_bucket_indices: list[list[int]] = []
-        self._base_bucket_weights: list[torch.Tensor] = []
-        self._active_bucket_indices: list[list[int]] = []
-        self._active_bucket_weights: list[torch.Tensor] = []
-
-        self.adaptive_stats = None
-        if self.config.adaptive_enabled:
-            self.adaptive_stats = AdaptiveBucketStats(
-                learning_signal_weight=self.config.adaptive_learning_signal_weight,
-                exploration_weight=self.config.adaptive_exploration_weight,
-                epsilon=self.config.epsilon,
-            )
-
-        self._cached_adaptive_probs: np.ndarray | None = None
-        self._last_adaptive_refresh_step = -1
-
-        self._build_bucket_index()
-
-    def _parse_metadata(self, example: dict[str, Any], index: int) -> _ParsedDifficultyMetadata:
-        difficulty_blob = _resolve_path(example, self.config.difficulty_field)
-        if not isinstance(difficulty_blob, dict):
-            return _ParsedDifficultyMetadata(
-                bucket_index=None,
-                bucket_count=None,
-                posterior_mean=None,
-                error=f"missing '{self.config.difficulty_field}' metadata for dataset index {index}",
-            )
-
-        bucket_index = _coerce_int(_resolve_path(difficulty_blob, self.config.bucket_index_field))
-        bucket_count = _coerce_int(_resolve_path(difficulty_blob, self.config.bucket_count_field))
-        posterior_mean = _coerce_float(_resolve_path(difficulty_blob, self.config.posterior_mean_field))
-
-        if bucket_index is None or bucket_index < 0:
-            return _ParsedDifficultyMetadata(
-                bucket_index=None,
-                bucket_count=bucket_count,
-                posterior_mean=posterior_mean,
-                error=f"invalid bucket_index for dataset index {index}",
-            )
-        if bucket_count is None or bucket_count <= 0:
-            return _ParsedDifficultyMetadata(
-                bucket_index=bucket_index,
-                bucket_count=None,
-                posterior_mean=posterior_mean,
-                error=f"invalid bucket_count for dataset index {index}",
-            )
-        if posterior_mean is None:
-            return _ParsedDifficultyMetadata(
-                bucket_index=bucket_index,
-                bucket_count=bucket_count,
-                posterior_mean=None,
-                error=f"invalid posterior_mean for dataset index {index}",
-            )
-        return _ParsedDifficultyMetadata(
-            bucket_index=bucket_index, bucket_count=bucket_count, posterior_mean=posterior_mean, error=None
-        )
-
-    def _build_bucket_index(self) -> None:
-        parsed_rows: list[_ParsedDifficultyMetadata] = []
-        observed_bucket_counts: set[int] = set()
-        max_bucket_index = -1
-
-        for dataset_index in range(len(self.dataset)):
-            parsed = self._parse_metadata(self.dataset[dataset_index], dataset_index)
-            if parsed.error is not None and self.config.strict_metadata:
-                raise ValueError(parsed.error)
-            if parsed.bucket_count is not None:
-                observed_bucket_counts.add(parsed.bucket_count)
-            if parsed.bucket_index is not None:
-                max_bucket_index = max(max_bucket_index, parsed.bucket_index)
-            parsed_rows.append(parsed)
-
-        if observed_bucket_counts:
-            if self.config.strict_metadata and len(observed_bucket_counts) > 1:
-                raise ValueError(
-                    f"inconsistent difficulty bucket_count values found: {sorted(observed_bucket_counts)}"
-                )
-            self.bucket_count = max(observed_bucket_counts)
-        elif max_bucket_index >= 0:
-            self.bucket_count = max_bucket_index + 1
-        else:
-            self.bucket_count = 1
-
-        self._base_bucket_indices = [[] for _ in range(self.bucket_count)]
-        bucket_weight_lists: list[list[float]] = [[] for _ in range(self.bucket_count)]
-        fallback_bucket = min(self.bucket_count - 1, self.bucket_count // 2)
-
-        for dataset_index, parsed in enumerate(parsed_rows):
-            if parsed.error is not None:
-                bucket_index = fallback_bucket
-                posterior_mean = _DEFAULT_POSTERIOR_MEAN
-                self.metadata_fallback_count += 1
-            else:
-                assert parsed.bucket_index is not None
-                bucket_index = int(np.clip(parsed.bucket_index, 0, self.bucket_count - 1))
-                posterior_mean = parsed.posterior_mean
-
-            if posterior_mean is None:
-                posterior_mean = _DEFAULT_POSTERIOR_MEAN
-            posterior_mean = float(np.clip(posterior_mean, 0.0, 1.0))
-            example_weight = self._compute_example_weight(posterior_mean)
-
-            self._index_to_bucket[dataset_index] = bucket_index
-            self._base_bucket_indices[bucket_index].append(dataset_index)
-            bucket_weight_lists[bucket_index].append(example_weight)
-
-        self._base_bucket_weights = [
-            torch.tensor(weight_list, dtype=torch.float64) for weight_list in bucket_weight_lists
-        ]
-        self._active_bucket_indices = [list(indices) for indices in self._base_bucket_indices]
-        self._active_bucket_weights = [weights.clone() for weights in self._base_bucket_weights]
-
-        if self.metadata_fallback_count > 0 and not self.config.strict_metadata:
-            logger.warning(
-                "Difficulty curriculum fell back to conservative defaults for %s/%s rows because metadata was missing "
-                "or invalid.",
-                self.metadata_fallback_count,
-                len(self.dataset),
-            )
-
-    def _compute_example_weight(self, posterior_mean: float) -> float:
-        probability = float(np.clip(posterior_mean, 0.0, 1.0))
-        uncertainty = 4.0 * probability * (1.0 - probability)
-        hardness = 1.0 - probability
-        return (
-            self.config.uncertainty_weight * uncertainty
-            + (1.0 - self.config.uncertainty_weight) * hardness
-            + self.config.epsilon
-        )
-
-    def __len__(self) -> int:
-        return self.num_samples
-
-    @property
-    def bucket_to_indices(self) -> tuple[tuple[int, ...], ...]:
-        return tuple(tuple(indices) for indices in self._base_bucket_indices)
-
-    def _get_current_step(self) -> int:
-        step = self.global_step_getter() if self.global_step_getter is not None else 0
-        return max(int(step), 0)
-
-    def get_progress(self, step: int | None = None) -> float:
-        if step is None:
-            step = self._get_current_step()
+    def get_progress(self, step: int) -> float:
         if step < self.config.warmup_steps:
             return 0.0
-        return min(1.0, (step - self.config.warmup_steps) / self.config.total_curriculum_steps)
+        return min(1.0, (step - self.config.warmup_steps) / self.config.total_steps)
 
-    def _smooth_progress(self, step: int | None = None) -> float:
+    def _smooth_progress(self, step: int) -> float:
         progress = self.get_progress(step)
         return progress * progress * (3.0 - 2.0 * progress)
 
     def _get_default_bucket_sigma(self) -> float:
         return max(0.85, 0.25 * max(self.bucket_count - 1, 1))
 
-    def _get_bucket_sigma(self, step: int | None = None) -> float:
+    def _get_bucket_sigma(self, step: int) -> float:
         sigma = self.config.bucket_sigma if self.config.bucket_sigma > 0 else self._get_default_bucket_sigma()
-        if step is not None and step < self.config.easy_focus_steps and self.config.easy_focus_sigma > 0:
-            return self.config.easy_focus_sigma
+        if step < self.config.bootstrap_steps and self.config.bootstrap_sigma > 0:
+            return self.config.bootstrap_sigma
         return sigma
 
     def _bucket_ratio_to_bucket_index(self, bucket_ratio: float) -> float:
         return float(self.bucket_count - 1) * bucket_ratio
 
-    def _get_target_bucket(self, step: int | None = None) -> float:
-        if step is None:
-            step = self._get_current_step()
+    def _get_target_bucket(self, step: int) -> float:
+        warmup_target_bucket = self._bucket_ratio_to_bucket_index(self.config.warmup_target)
+        final_target_bucket = self._bucket_ratio_to_bucket_index(self.config.final_target)
 
-        warmup_target_bucket = self._bucket_ratio_to_bucket_index(self.config.warmup_target_bucket_ratio)
-        final_target_bucket = self._bucket_ratio_to_bucket_index(self.config.final_target_bucket_ratio)
-
-        if self.config.easy_focus_steps > 0 and step < self.config.easy_focus_steps:
-            easy_progress = min(1.0, step / self.config.easy_focus_steps)
-            bootstrap_target_bucket = self._bucket_ratio_to_bucket_index(self.config.bootstrap_target_bucket_ratio)
-            return bootstrap_target_bucket + (warmup_target_bucket - bootstrap_target_bucket) * easy_progress
+        if self.config.bootstrap_steps > 0 and step < self.config.bootstrap_steps:
+            bootstrap_progress = min(1.0, step / self.config.bootstrap_steps)
+            bootstrap_target_bucket = self._bucket_ratio_to_bucket_index(self.config.bootstrap_target)
+            return bootstrap_target_bucket + (warmup_target_bucket - bootstrap_target_bucket) * bootstrap_progress
 
         smooth_progress = self._smooth_progress(step)
         return warmup_target_bucket + (final_target_bucket - warmup_target_bucket) * smooth_progress
 
-    def _available_bucket_mask(self) -> np.ndarray:
-        return np.array([1.0 if indices else 0.0 for indices in self._active_bucket_indices], dtype=np.float64)
-
-    def get_static_bucket_probs(self, step: int | None = None) -> np.ndarray:
+    def build_probs(self, step: int, available_mask: np.ndarray) -> np.ndarray:
         if self.bucket_count == 1:
             return np.ones(1, dtype=np.float64)
-
-        if step is None:
-            step = self._get_current_step()
 
         smooth_progress = self._smooth_progress(step)
         target_bucket = self._get_target_bucket(step)
@@ -468,22 +392,215 @@ class BetaBinomialDifficultySampler(Sampler[int]):
         static_probs[:-1] = (1.0 - hard_bucket_frac) * non_hard_probs
         static_probs[-1] = hard_bucket_frac
 
-        mask = self._available_bucket_mask()
-        static_probs *= mask
-        if mask.sum() == 0:
+        static_probs *= available_mask
+        if available_mask.sum() == 0:
             return np.ones(self.bucket_count, dtype=np.float64) / self.bucket_count
         if static_probs.sum() <= 0:
-            return _normalize_probs(mask)
+            return _normalize_probs(available_mask)
         return _normalize_probs(static_probs)
 
+
+def _parse_difficulty_metadata(
+    example: dict[str, Any], index: int, metadata_config: DifficultyCurriculumMetadataConfig
+) -> _ParsedDifficultyMetadata:
+    difficulty_blob = _resolve_path(example, metadata_config.field)
+    if not isinstance(difficulty_blob, dict):
+        return _ParsedDifficultyMetadata(
+            bucket_index=None,
+            bucket_count=None,
+            posterior_mean=None,
+            error=f"missing '{metadata_config.field}' metadata for dataset index {index}",
+        )
+
+    bucket_index = _coerce_int(_resolve_path(difficulty_blob, metadata_config.bucket_index_field))
+    bucket_count = _coerce_int(_resolve_path(difficulty_blob, metadata_config.bucket_count_field))
+    posterior_mean = _coerce_float(_resolve_path(difficulty_blob, metadata_config.posterior_mean_field))
+
+    if bucket_index is None or bucket_index < 0:
+        return _ParsedDifficultyMetadata(
+            bucket_index=None,
+            bucket_count=bucket_count,
+            posterior_mean=posterior_mean,
+            error=f"invalid bucket_index for dataset index {index}",
+        )
+    if bucket_count is None or bucket_count <= 0:
+        return _ParsedDifficultyMetadata(
+            bucket_index=bucket_index,
+            bucket_count=None,
+            posterior_mean=posterior_mean,
+            error=f"invalid bucket_count for dataset index {index}",
+        )
+    if posterior_mean is None:
+        return _ParsedDifficultyMetadata(
+            bucket_index=bucket_index,
+            bucket_count=bucket_count,
+            posterior_mean=None,
+            error=f"invalid posterior_mean for dataset index {index}",
+        )
+    return _ParsedDifficultyMetadata(
+        bucket_index=bucket_index, bucket_count=bucket_count, posterior_mean=posterior_mean, error=None
+    )
+
+
+def _compute_example_weight(posterior_mean: float, uncertainty_weight: float, epsilon: float) -> float:
+    probability = float(np.clip(posterior_mean, 0.0, 1.0))
+    uncertainty = 4.0 * probability * (1.0 - probability)
+    hardness = 1.0 - probability
+    return uncertainty_weight * uncertainty + (1.0 - uncertainty_weight) * hardness + epsilon
+
+
+def _build_difficulty_bucket_index(
+    dataset, metadata_config: DifficultyCurriculumMetadataConfig, uncertainty_weight: float, epsilon: float
+) -> _DifficultyBucketIndex:
+    parsed_rows: list[_ParsedDifficultyMetadata] = []
+    observed_bucket_counts: set[int] = set()
+    max_bucket_index = -1
+
+    for dataset_index in range(len(dataset)):
+        parsed = _parse_difficulty_metadata(dataset[dataset_index], dataset_index, metadata_config)
+        if parsed.error is not None and metadata_config.strict:
+            raise ValueError(parsed.error)
+        if parsed.bucket_count is not None:
+            observed_bucket_counts.add(parsed.bucket_count)
+        if parsed.bucket_index is not None:
+            max_bucket_index = max(max_bucket_index, parsed.bucket_index)
+        parsed_rows.append(parsed)
+
+    if observed_bucket_counts:
+        if metadata_config.strict and len(observed_bucket_counts) > 1:
+            raise ValueError(f"inconsistent difficulty bucket_count values found: {sorted(observed_bucket_counts)}")
+        bucket_count = max(observed_bucket_counts)
+    elif max_bucket_index >= 0:
+        bucket_count = max_bucket_index + 1
+    else:
+        bucket_count = 1
+
+    bucket_to_indices: list[list[int]] = [[] for _ in range(bucket_count)]
+    bucket_weight_lists: list[list[float]] = [[] for _ in range(bucket_count)]
+    index_to_bucket: dict[int, int] = {}
+    metadata_fallback_count = 0
+    fallback_bucket = min(bucket_count - 1, bucket_count // 2)
+
+    for dataset_index, parsed in enumerate(parsed_rows):
+        if parsed.error is not None:
+            bucket_index = fallback_bucket
+            posterior_mean = _DEFAULT_POSTERIOR_MEAN
+            metadata_fallback_count += 1
+        else:
+            assert parsed.bucket_index is not None
+            bucket_index = int(np.clip(parsed.bucket_index, 0, bucket_count - 1))
+            posterior_mean = parsed.posterior_mean
+
+        if posterior_mean is None:
+            posterior_mean = _DEFAULT_POSTERIOR_MEAN
+        example_weight = _compute_example_weight(
+            posterior_mean=float(np.clip(posterior_mean, 0.0, 1.0)),
+            uncertainty_weight=uncertainty_weight,
+            epsilon=epsilon,
+        )
+
+        index_to_bucket[dataset_index] = bucket_index
+        bucket_to_indices[bucket_index].append(dataset_index)
+        bucket_weight_lists[bucket_index].append(example_weight)
+
+    return _DifficultyBucketIndex(
+        index_to_bucket=index_to_bucket,
+        bucket_to_indices=tuple(tuple(indices) for indices in bucket_to_indices),
+        bucket_weights=tuple(torch.tensor(weight_list, dtype=torch.float64) for weight_list in bucket_weight_lists),
+        bucket_count=bucket_count,
+        metadata_fallback_count=metadata_fallback_count,
+    )
+
+
+class DifficultyCurriculumSampler(Sampler[int]):
+    """Bucket-aware curriculum sampler for difficulty-annotated prompt datasets."""
+
+    def __init__(
+        self,
+        dataset,
+        num_samples: int,
+        config: DifficultyCurriculumConfig,
+        global_step_getter: Callable[[], int] | None,
+    ) -> None:
+        if num_samples <= 0:
+            raise ValueError("num_samples must be > 0")
+
+        self.dataset = dataset
+        self.num_samples = num_samples
+        self.config = config
+        self.global_step_getter = global_step_getter
+
+        self._generator = torch.Generator()
+        self._generator.manual_seed(self.config.seed)
+
+        bucket_index = _build_difficulty_bucket_index(
+            dataset=dataset,
+            metadata_config=self.config.metadata,
+            uncertainty_weight=self.config.uncertainty_weight,
+            epsilon=self.config.epsilon,
+        )
+        self._index_to_bucket = dict(bucket_index.index_to_bucket)
+        self.bucket_count = bucket_index.bucket_count
+        self.metadata_fallback_count = bucket_index.metadata_fallback_count
+        self._schedule = _DifficultyCurriculumSchedule(self.config.schedule, self.bucket_count)
+
+        self._excluded_indices: set[int] = set()
+        self._base_bucket_indices = [list(indices) for indices in bucket_index.bucket_to_indices]
+        self._base_bucket_weights = [weights.clone() for weights in bucket_index.bucket_weights]
+        self._active_bucket_indices = [list(indices) for indices in self._base_bucket_indices]
+        self._active_bucket_weights = [weights.clone() for weights in self._base_bucket_weights]
+
+        self.adaptive_stats = None
+        if self.config.adaptive.enabled:
+            self.adaptive_stats = AdaptiveBucketStats(
+                learning_signal_weight=self.config.adaptive.learning_weight,
+                exploration_weight=self.config.adaptive.exploration_weight,
+                epsilon=self.config.epsilon,
+            )
+
+        self._cached_adaptive_probs: np.ndarray | None = None
+        self._last_adaptive_refresh_step = -1
+
+        if self.metadata_fallback_count > 0 and not self.config.metadata.strict:
+            logger.warning(
+                "Difficulty curriculum fell back to conservative defaults for %s/%s rows because metadata was missing "
+                "or invalid.",
+                self.metadata_fallback_count,
+                len(self.dataset),
+            )
+
+    def __len__(self) -> int:
+        return self.num_samples
+
+    @property
+    def bucket_to_indices(self) -> tuple[tuple[int, ...], ...]:
+        return tuple(tuple(indices) for indices in self._base_bucket_indices)
+
+    def _get_current_step(self) -> int:
+        step = self.global_step_getter() if self.global_step_getter is not None else 0
+        return max(int(step), 0)
+
+    def get_progress(self, step: int | None = None) -> float:
+        if step is None:
+            step = self._get_current_step()
+        return self._schedule.get_progress(step)
+
+    def _available_bucket_mask(self) -> np.ndarray:
+        return np.array([1.0 if indices else 0.0 for indices in self._active_bucket_indices], dtype=np.float64)
+
+    def get_static_bucket_probs(self, step: int | None = None) -> np.ndarray:
+        if step is None:
+            step = self._get_current_step()
+        return self._schedule.build_probs(step, self._available_bucket_mask())
+
     def get_adaptive_bucket_probs(self, step: int | None = None) -> np.ndarray | None:
-        if not self.config.adaptive_enabled or self.adaptive_stats is None or self.adaptive_stats.total_count == 0:
+        if not self.config.adaptive.enabled or self.adaptive_stats is None or self.adaptive_stats.total_count == 0:
             return None
 
         refresh_step = self._get_current_step() if step is None else step
         if (
             self._cached_adaptive_probs is not None
-            and refresh_step - self._last_adaptive_refresh_step < self.config.adaptive_update_every
+            and refresh_step - self._last_adaptive_refresh_step < self.config.adaptive.update_every
         ):
             return self._cached_adaptive_probs.copy()
 
@@ -502,9 +619,7 @@ class BetaBinomialDifficultySampler(Sampler[int]):
         if adaptive_probs is None:
             return static_probs
 
-        final_probs = (
-            1.0 - self.config.adaptive_blend_weight
-        ) * static_probs + self.config.adaptive_blend_weight * adaptive_probs
+        final_probs = (1.0 - self.config.adaptive.blend) * static_probs + self.config.adaptive.blend * adaptive_probs
         final_probs *= self._available_bucket_mask()
         if final_probs.sum() <= 0:
             return static_probs
@@ -578,7 +693,7 @@ class BetaBinomialDifficultySampler(Sampler[int]):
         rewards: list[float] | np.ndarray,
         advantages: list[float] | np.ndarray | None = None,
     ) -> None:
-        if not self.config.adaptive_enabled or self.adaptive_stats is None:
+        if not self.config.adaptive.enabled or self.adaptive_stats is None:
             return
         if len(dataset_indices) == 0:
             return
@@ -595,7 +710,7 @@ class BetaBinomialDifficultySampler(Sampler[int]):
             metrics[f"curriculum/static_bucket_prob_{bucket_index}"] = float(probability)
 
         adaptive_probs = self.get_adaptive_bucket_probs(step)
-        if self.config.adaptive_enabled:
+        if self.config.adaptive.enabled:
             if adaptive_probs is None:
                 adaptive_probs = np.zeros(self.bucket_count, dtype=np.float64)
             for bucket_index, probability in enumerate(adaptive_probs):
@@ -611,7 +726,7 @@ class BetaBinomialDifficultySampler(Sampler[int]):
         for bucket_index, count in enumerate(sampled_counts):
             metrics[f"curriculum/sampled_bucket_count_{bucket_index}"] = float(count)
 
-        if self.config.adaptive_enabled and self.adaptive_stats is not None:
+        if self.config.adaptive.enabled and self.adaptive_stats is not None:
             for bucket_index in range(self.bucket_count):
                 metrics[f"curriculum/bucket_reward_mean_{bucket_index}"] = float(
                     self.adaptive_stats.get_mean_reward(bucket_index)
